audio-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai-whisper==20231117
numpy==2.1.3
scipy==1.14.1
matplotlib==3.9.2